HZN-973: Automatically add the according exception mapper to the osgi rest service
