+++ features/opennms-osgi-core-rest/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,17 @@
+<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.1.0"
+           xsi:schemaLocation="
+http://www.osgi.org/xmlns/blueprint/v1.0.0 http://www.osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.1.0 http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.1.0.xsd
+http://www.osgi.org/xmlns/blueprint-ext/v1.1.0 http://aries.apache.org/schemas/blueprint-ext/blueprint-ext-1.1.xsd">
+
+    <service interface="javax.ws.rs.ext.ExceptionMapper">
+        <bean class="org.opennms.web.rest.support.NotFoundProvider"/>
+    </service>
+    <service interface="javax.ws.rs.ext.ExceptionMapper">
+        <bean class="org.opennms.web.rest.support.NoSuchElementProvider"/>
+    </service>
+    <service interface="javax.ws.rs.ext.ExceptionMapper">
+        <bean class="org.opennms.web.rest.support.ErrorResponseProvider"/>
+    </service>
+</blueprint>
