+++ features/pom.xml
@@ -60,6 +60,9 @@
     <!-- OpenNMS REST -->
     <module>rest</module>
 
+    <!-- OpenNMS rest provider module -->
+    <module>rest-provider</module>
+
     <!-- OpenNMS OSGI Core -->
     <module>opennms-osgi-core</module>
