+++ features/rest-provider/src/main/java/org/opennms/web/rest/support/NoSuchElementProvider.java
