+++ opennms-assemblies/webapp-full/pom.xml
@@ -150,6 +150,17 @@
         </exclusion>
       </exclusions>
     </dependency>
+    <dependency>
+      <groupId>org.opennms.features</groupId>
+      <artifactId>org.opennms.features.rest-provider</artifactId>
+      <version>${project.version}</version>
+      <exclusions>
+        <exclusion>
+          <groupId>*</groupId>
+          <artifactId>*</artifactId>
+        </exclusion>
+      </exclusions>
+    </dependency>
     <!-- The bsm.rest.api uses org.reflections, manually copy it to the web app -->
     <dependency>
       <groupId>org.reflections</groupId>
